rawler

Module decoders

Re-exports

pub use super::rawimage::*;

Modules

ari

arw

cr2

cr3

crw

dcr

dcs

dng

erf

iiq

kdc

mef

mos

mrw

nef

nkd

nrw

orf

pef

qtk

raf

rw2

srw

tfr

x3f

Structs

Camera

Contains sanitized information about the raw image’s properties

RawDecodeParams

RawLoader

The struct that holds all the info about the cameras and is able to decode a file

RawMetadata

Enums

FormatHint

Orientation

Possible orientations of an image

WellKnownIFD

Statics

BUG

CAMERAS_TOML

SAMPLE

Traits

Decoder

Readable

Functions

decode_threaded

decode_threaded_chunked

This is used for streams where not chunked at line boundaries.

decode_threaded_multiline

decode_unthreaded

ok_cfa_image

ok_cfa_image_with_black_white

ok_cfa_image_with_blacklevels

supported_extensions

Get list of supported file extensions. All names are upper-case.

Type Aliases

ReadableBoxed